+++ bgn_ngx/bgn_src/clist.c
@@ -1183,7 +1183,7 @@ EC_BOOL clist_loop(CLIST *clist,
     if(MAX_NUM_OF_FUNC_PARAS < func_para_num)
     {
         dbg_log(SEC_0044_CLIST, 0)(LOGSTDOUT, "error:clist_loop: func_para_num %ld overflow which must be smaller than %ld\n",
-                           func_para_num, MAX_NUM_OF_FUNC_PARAS);
+                           func_para_num, (UINT32)MAX_NUM_OF_FUNC_PARAS);
         return (EC_FALSE);
     }
 
@@ -1958,7 +1958,7 @@ EC_BOOL clist_loop_no_lock(CLIST *clist,
     if(MAX_NUM_OF_FUNC_PARAS < func_para_num)
     {
         dbg_log(SEC_0044_CLIST, 0)(LOGSTDOUT, "error:clist_loop_no_lock: func_para_num %ld overflow which must be smaller than %ld\n",
-                           func_para_num, MAX_NUM_OF_FUNC_PARAS);
+                           func_para_num, (UINT32)MAX_NUM_OF_FUNC_PARAS);
         return (EC_FALSE);
     }
